replace chunks bool array, with bit fields.
sizeof(bool) == 1. This means that for small buckets, where there is
alot of chunks we add alot of extra overhead, as we store one bit as 8
bit variable. This PR changes bool chunks[] to uint64_t chunks[], so
we can tightly pack bits, and reduce overhead. This might also have
positive impact on the performance, do to reducting memory
overhead (cpu cache),
using performance effective intrinsic to free chunk, instead of
iteration over the array.
In this PR i also replaced added union to store bitmap in side of the
slab struct, if there is less then 64 chunks. This remove need for
extra allocation from global alloc. Global alloc is "slooooow", so
this should be a performance improvement. I whould propose to
consider, reducting bitmap size to 64 bites, but this needs more
consideration as slab size is configured by the user.

Union name and it's fields are not final - you can rename them.

Signed-off-by: Łukasz Plewa <lukasz.plewa@intel.com>